requirements.txt
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
idna==3.10
prettytable==3.16.0
python-dotenv==1.1.0
requests==2.32.3
soupsieve==2.6
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
